cleanup schema
