+++ cspell.config.json
@@ -3,7 +3,11 @@
     ".github/**",
     "node_modules/**",
     "dist/**",
+    "compiled/",
     "yarn.lock",
+    "staticRegistry.ts",
+    "*.schema.json",
+    "*.d.ts",
     ".yarn/**",
     ".yarnrc.yml",
     ".gitignore",
@@ -185,7 +189,12 @@
     "comparitor",
     "pangea",
     "attw",
+    "keyvalue",
+    "creds",
+    "traceql",
+    "logql",
     "packagejson",
+    "marcbachmann",
     "publint",
     "uplot",
     "arethetypeswrong"
