+++ package.json
@@ -80,6 +80,7 @@
     "@grafana/tsconfig": "^2.0.0",
     "@grafana/ui": "^10.4.0",
     "@rollup/plugin-commonjs": "^29.0.0",
+    "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^16.0.0",
     "@storybook/addon-docs": "10.2.5",
     "@storybook/addon-links": "10.2.5",
@@ -94,6 +95,7 @@
     "@types/chance": "^1.1.0",
     "@types/lodash": "^4.14.194",
     "@types/memoize-one": "^5.1.2",
+    "@types/node": "^25.9.3",
     "@types/react": "18.3.28",
     "@types/react-calendar": "^3.1.2",
     "@types/testing-library__jest-dom": "^5.9.5",
