+++ src/index.ts
@@ -5,3 +5,4 @@
  */
 export * from './components';
 export { SqlDatasource } from './datasource/SqlDatasource';
+export * from './schema';
